requirements.txt
anyjson==0.3.3
Django==1.9.7
django-cors-headers==1.1.0
django-debug-toolbar==1.7
django-user-sessions==1.3.0
djangorestframework==3.4.6
gunicorn==19.5.0
Markdown==2.6.6
msgpack-python==0.4.7
phonenumbers==7.4.0
requests==2.11.1
psycopg2==2.7.1
rehive==1.1.0
django-filter==1.1.0
